+++ pkg/mime/parser_test.go
@@ -90,7 +90,7 @@ Mime-Version: 1.0
 Subject: Encrypted Message
 Return-Path: <xyz@xyz.com>
 From: XYZ <xyz@xyz.com>
-X-Pm-Conversationid-Id: gNX9bDPLmBgFZ-C3Tdlb628cas1Xl0m4dql5nsWzQAEI-WQv0ytfwPR4-PWELEK0_87XuFOgetc239Y0pjPYHQ==
+X-Pm-ConversationID-Id: gNX9bDPLmBgFZ-C3Tdlb628cas1Xl0m4dql5nsWzQAEI-WQv0ytfwPR4-PWELEK0_87XuFOgetc239Y0pjPYHQ==
 X-Pm-Date: Sun, 10 Mar 2019 18:10:06 +0100
 Message-Id: <68c11e46-e611-d9e4-edc1-5ec96bac77cc@unicoderns.com>
 X-Pm-Transfer-Encryption: TLSv1.2 with cipher ECDHE-RSA-AES256-GCM-SHA384 (256/256 bits)