+++ Changelog.md
@@ -2,6 +2,11 @@
 
 Changelog [format](http://keepachangelog.com/en/1.0.0/)
 
+## unreleased
+
+### Changed
+* GODT-308 better user error message when request is canceled
+
 ## [v1.2.7] Donghai-hotfix - beta (2020-05-07)
 
 ### Added